+++ Subliminal/guessit/containers.py
@@ -38,7 +38,7 @@ def _get_span(prop, match):
                 start = span[0]
             if end is None or span[1] > end:
                 end = span[1]
-        return (start, end)
+        return start, end
     else:
         return match.span()
         start = span[0]
@@ -50,13 +50,13 @@ def _trim_span(span, value, blanks = sep):
 
     for i in range(0, len(value)):
         if value[i] in blanks:
-            start = start + 1
+            start += 1
         else:
             break
 
     for i in reversed(range(0, len(value))):
         if value[i] in blanks:
-            end = end - 1
+            end -= 1
         else:
             break
     if end <= start:
@@ -186,38 +186,58 @@ class FormatterValidator(object):
             return formatted
 
 
+def _get_positions(prop, string, node, match, entry_start, entry_end):
+    span = match.span()
+    start = span[0]
+    end = span[1]
+
+    at_start = True
+    at_end = True
+
+    while start > 0:
+        start -= 1
+        if string[start] not in sep:
+            at_start = False
+            break
+    while end < len(string) - 1:
+        end += 1
+        if string[end] not in sep:
+            at_end = False
+            break
+    return at_start, at_end
+
+
 class WeakValidator(DefaultValidator):
     """Make sure our match is surrounded by separators and is the first or last element in the string"""
     def validate(self, prop, string, node, match, entry_start, entry_end):
         if super(WeakValidator, self).validate(prop, string, node, match, entry_start, entry_end):
-            span = match.span()
-            start = span[0]
-            end = span[1]
+            at_start, at_end = _get_positions(prop, string, node, match, entry_start, entry_end)
+            return at_start or at_end
+        return False
 
-            at_start = True
-            at_end = True
 
-            while start > 0:
-                start = start - 1
-                if string[start] not in sep:
-                    at_start = False
-                    break
-            if at_start:
+class NeighborValidator(DefaultValidator):
+    """Make sure the node is next another one"""
+    def validate(self, prop, string, node, match, entry_start, entry_end):
+        at_start, at_end = _get_positions(prop, string, node, match, entry_start, entry_end)
+
+        if at_start:
+            previous_leaf = node.root.previous_leaf(node)
+            if previous_leaf is not None:
                 return True
-            while end < len(string) - 1:
-                end = end + 1
-                if string[end] not in sep:
-                    at_end = False
-                    break
-            if at_end:
+
+        if at_end:
+            next_leaf = node.root.next_leaf(node)
+            if next_leaf is not None:
                 return True
+
         return False
 
 
 class LeavesValidator(DefaultValidator):
     def __init__(self, lambdas=None, previous_lambdas=None, next_lambdas=None, both_side=False, default_=True):
-        self.previous_lambdas = previous_lambdas if not previous_lambdas is None else []
-        self.next_lambdas = next_lambdas if not next_lambdas is None else []
+        self.previous_lambdas = previous_lambdas if previous_lambdas is not None else []
+        self.next_lambdas = next_lambdas if next_lambdas is not None else []
         if lambdas:
             self.previous_lambdas.extend(lambdas)
             self.next_lambdas.extend(lambdas)
@@ -248,7 +268,7 @@ class LeavesValidator(DefaultValidator):
             for leaf in node.root.previous_leaves(node):
                 for lambda_ in self.previous_lambdas:
                     ret = self._check_rule(lambda_, leaf)
-                    if not ret is None:
+                    if ret is not None:
                         return ret
             return False
 
@@ -257,7 +277,7 @@ class LeavesValidator(DefaultValidator):
             for leaf in node.root.next_leaves(node):
                 for lambda_ in self.next_lambdas:
                     ret = self._check_rule(lambda_, leaf)
-                    if not ret is None:
+                    if ret is not None:
                         return ret
             return False
 
@@ -294,7 +314,7 @@ class _Property:
         else:
             self.keys = []
         self.canonical_form = canonical_form
-        if not pattern is None:
+        if pattern is not None:
             self.pattern = pattern
         else:
             self.pattern = canonical_form
@@ -323,13 +343,13 @@ class _Property:
         formatter = None
         if isinstance(self.formatter, dict):
             formatter = self.formatter.get(group_name)
-            if formatter is None and not group_name is None:
+            if formatter is None and group_name is not None:
                 formatter = self.formatter.get(None)
         else:
             formatter = self.formatter
         if isinstance(formatter, types.FunctionType):
             return formatter(value)
-        elif not formatter is None:
+        elif formatter is not None:
             return formatter.format(value)
         return value
 
@@ -520,7 +540,7 @@ class PropertiesContainer(object):
             entries_dict = {}
             for entry in entries:
                 for key in prop.keys:
-                    if not key in entries_dict:
+                    if key not in entries_dict:
                         entries_dict[key] = []
                     entries_dict[key].append(entry)
 
@@ -552,18 +572,17 @@ class PropertiesContainer(object):
             def _sorting(x):
                 _, x_match = x
                 x_start, x_end = x_match.span()
-                return (x_start - x_end)
+                return x_start - x_end
 
             ret.sort(key=_sorting)
 
         return ret
 
-    def as_guess(self, found_properties, input=None, filter=None, sep_replacement=None, multiple=False, *args, **kwargs):
-        if filter is None:
-            filter = lambda property, *args, **kwargs: True
+    def as_guess(self, found_properties, input=None, filter_=None, sep_replacement=None, multiple=False, *args, **kwargs):
+        if filter_ is None:
+            filter_ = lambda property, *args, **kwargs: True
         guesses = [] if multiple else None
-        for property in found_properties:
-            prop, match = property
+        for prop, match in found_properties:
             first_key = None
             for key in prop.keys:
                 # First property key will be used as base for effective name
@@ -596,7 +615,7 @@ class PropertiesContainer(object):
                                     guess[name] = value
                             if group_name:
                                 guess.metadata(prop).span = match.span(group_name)
-            if filter(guess):
+            if filter_(guess):
                 if multiple:
                     guesses.append(guess)
                 else:
@@ -609,7 +628,7 @@ class PropertiesContainer(object):
         if input is None:
             return None
         value = input
-        if not span is None:
+        if span is not None:
             value = value[span[0]:span[1]]
         value = input[span[0]:span[1]] if input else None
         if sep_replacement:
@@ -679,7 +698,7 @@ class QualitiesContainer():
                 del self._qualities[name]
         else:
             property_qualities = self._qualities.get(name)
-            if not property_qualities is None:
+            if property_qualities is not None:
                 for property_canonical_form in canonical_forms:
                     if property_canonical_form in property_qualities:
                         del property_qualities[property_canonical_form]
@@ -708,7 +727,7 @@ class QualitiesContainer():
         for prop in props:
             prop_value = guess.get(prop)
             prop_qualities = self._qualities.get(prop)
-            if not prop_value is None and not prop_qualities is None:
+            if prop_value is not None and prop_qualities is not None:
                 rate += prop_qualities.get(prop_value, 0)
         return rate