+++ Subliminal/guessit/date.py
@@ -31,13 +31,14 @@ _dsep = r'[-/ \.]'
 _dsep_bis = r'[-/ \.x]'
 
 date_regexps = [
-    re.compile('[^\d](\d{8})[^\d]', re.IGNORECASE),
-    re.compile('[^\d](\d{6})[^\d]', re.IGNORECASE),
+    re.compile('%s(\d{8})%s' % (_dsep, _dsep), re.IGNORECASE),
+    re.compile('%s(\d{6})%s' % (_dsep, _dsep), re.IGNORECASE),
     re.compile('[^\d](\d{2})%s(\d{1,2})%s(\d{1,2})[^\d]' % (_dsep, _dsep), re.IGNORECASE),
     re.compile('[^\d](\d{1,2})%s(\d{1,2})%s(\d{2})[^\d]' % (_dsep, _dsep), re.IGNORECASE),
     re.compile('[^\d](\d{4})%s(\d{1,2})%s(\d{1,2})[^\d]' % (_dsep_bis, _dsep), re.IGNORECASE),
     re.compile('[^\d](\d{1,2})%s(\d{1,2})%s(\d{4})[^\d]' % (_dsep, _dsep_bis), re.IGNORECASE),
     re.compile('[^\d](\d{1,2}(?:st|nd|rd|th)?%s(?:[a-z]{3,10})%s\d{4})[^\d]' % (_dsep, _dsep), re.IGNORECASE)]
+
 
 def valid_year(year, today=None):
     """Check if number is a valid year"""
@@ -104,7 +105,6 @@ def search_date(string, year_first=None, day_first=True):
         return None, None
 
     today = datetime.date.today()
-    date = None
 
     # If day_first/year_first is undefined, parse is made using both possible values.
     yearfirst_opts = [False, True]
@@ -124,6 +124,6 @@ def search_date(string, year_first=None, day_first=True):
             pass
         # check date plausibility
         if date and valid_year(date.year, today=today):
-            return (date.date(), (start+1, end-1)) #compensate for sentinels
+            return date.date(), (start+1, end-1) #compensate for sentinels
 
     return None, None