+++ Subliminal/guessit/language.py
@@ -169,7 +169,7 @@ LNG_COMMON_WORDS = frozenset([
     'no', 'non', 'war', 'min', 'new', 'car', 'day', 'bad', 'bat', 'fan',
     'fry', 'cop', 'zen', 'gay', 'fat', 'one', 'cherokee', 'got', 'an', 'as',
     'cat', 'her', 'be', 'hat', 'sun', 'may', 'my', 'mr', 'rum', 'pi', 'bb', 'bt',
-    'tv', 'aw', 'by', 'md', 'mp', 'cd', 'lt', 'gt'
+    'tv', 'aw', 'by', 'md', 'mp', 'cd', 'lt', 'gt', 'in', 'ad', 'ice', 'ay', 'at',
     # french words
     'bas', 'de', 'le', 'son', 'ne', 'ca', 'ce', 'et', 'que',
     'mal', 'est', 'vol', 'or', 'mon', 'se', 'je', 'tu', 'me',
@@ -234,7 +234,7 @@ def find_possible_languages(string, allowed_languages=None):
         for prefix in lang_prefixes:
             if lang_word.startswith(prefix):
                 lang_word = lang_word[len(prefix):]
-        if not lang_word in common_words:
+        if lang_word not in common_words:
             try:
                 lang = Language.fromguessit(lang_word)
                 if allowed_languages:
@@ -274,8 +274,8 @@ def search_language(string, allowed_languages=None):
 
         # only allow those languages that have a 2-letter code, those that
         # don't are too esoteric and probably false matches
-        #if language.lang not in lng3_to_lng2:
-        #    continue
+        # if language.lang not in lng3_to_lng2:
+        #     continue
 
         # confidence depends on alpha2, alpha3, english name, ...
         if len(lang) == 2: