+++ Subliminal_Licensing/python-guessit-0.10.3/HISTORY.rst
@@ -1,14 +1,41 @@
-.. This is your project NEWS file which will contain the release notes.
-.. Example: http://www.python.org/download/releases/2.6/NEWS.txt
-.. The content of this file, along with README.rst, will appear in your
-.. project's PyPI page.
-
 History
 =======
 
-0.9.4 (unreleased)
+0.10.3 (2015-04-04)
 ------------------
-* Nothing changed yet.
+* Fix issues related to unicode encoding/decoding
+* Fix possible crashes in guess_video_rexps
+* Fix invalid guess result when crc32 contains 6 digits than can be parsed as a date
+
+0.10.2 (2015-03-08)
+------------------
+* Use common words to resolve conflicts on strings
+* Bump babelfish version
+* Fix setuptools deprecation warning
+* Package argparse dependency only if python<2.7
+
+0.10.1 (2015-01-05)
+------------------
+* Avoid word Stay to be recognized as AY subtitle
+* Fixed exception when no unidentified leaves remains
+* Avoid usage of deprecated EntryPoint.load() require argument
+* Fixed invalid raw data for some properties (title, series and maybe others)
+
+
+0.10.0 (2014-12-27)
+------------------
+* Fixed exception when serie title starts with Ep
+* Fixed exception when trying to parse a full length country name
+* Removed deprecated optparse module, replaced by argparse
+
+
+0.9.4 (2014-11-10)
+------------------
+* Fixed exception when filename contains multiple languages ISO codes
+* Fixed transformers initialization logging
+* Fixed possible exception in language transformer
+* Added more words to common english words
+
 
 0.9.3 (2014-09-14)
 ------------------
@@ -17,6 +44,7 @@ History
 * Better detection of ``format``` = ``BluRay`` (when followed by Rip)
 * Recognize ``RC`` as ``R5``
 * Recognize ``WEB-HD```and ``ẀEB`` as ``WEB-DL``
+
 
 0.9.2 (2014-09-13)
 ------------------
@@ -27,6 +55,7 @@ History
 * Fixed usage not showing when invalid options were passed
 * Added ``PAL``, ``SECAM`` and ``NTSC`` to ``other`` possible values
 * Recognize DVD-9 and DVD-5 as ``format`` = ``DVD`` property
+
 
 0.9.1 (2014-09-06)
 ------------------
@@ -39,6 +68,7 @@ History
 * Better guessing of ``format`` = ``DVD`` property (DVD-R pattern)
 * Some ``info`` logs changed to ``debug`` for quiet integration
 * Small fixes
+
 
 0.9.0 (2014-09-05)
 ------------------